Moved user stats to the services module
