+++ scoring/views.py
@@ -3,9 +3,9 @@ from django.db.models import Q, Count
 from django.views.decorators.cache import cache_page
 from rest_framework import status, serializers
 from rest_framework.decorators import api_view, permission_classes
+from rest_framework.exceptions import NotFound
 from rest_framework.generics import get_object_or_404
 from rest_framework.permissions import AllowAny
-from rest_framework.exceptions import NotFound
 from rest_framework.request import Request
 from rest_framework.response import Response
 
@@ -23,9 +23,8 @@ from scoring.serializers import (
     GetLeaderboardSerializer,
 )
 from scoring.utils import get_contributions, update_project_leaderboard
-
 from users.models import User
-from users.views import serialize_profile
+from users.services.profile_stats import serialize_profile
 
 
 @api_view(["GET"])
